requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.0.3
pillow==10.1.0
httpx==0.25.2
requests==2.31.0
python-dateutil==2.8.2
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
gunicorn==21.2.0
prometheus-client==0.19.0
python-slugify==8.0.1
pytz==2023.3